added insert, delete queries and modyfied Card table

+++ src/main/java/model/Entities/Rental.java
@@ -17,7 +17,7 @@ public class Rental {
     int startOdometer;
 
     //Card for which rental was made
-    int CardNo;
+    long CardNo;
     String cardName;
     int expDate;
